linux: Main fourth pass -- 131 -> 85 unique errors
 - Pointer-to-int casts go through intptr_t. Most are hash/sort keys where the
   hash VALUE may narrow (GDecal.h, GRenderCore.h, GSceneInternal.h, basic2.h);
   aiHeight.cpp's pointer-identity map and scriptCommon.cpp's lua thread id
   widen to intptr_t so the KEY stays lossless. UIWindow.cpp's wild-pointer
   guard keeps its 32-bit upper bound only where a pointer is 4 bytes.
 - GParticleFormat.cpp: the relocation pass is made explicit but the format is
   NOT fixed -- SParticle is overlaid on the file image with pointer-sized
   `keys` fields that the Windows content tools wrote as 32-bit offsets, so on
   a 64-bit build the struct stride disagrees with the file. It compiles now;
   loading particles on 64-bit needs the on-disk struct to stop containing
   pointers. Called out in a comment at the site.
 - Calls into a dependent base qualified with this-> (Sync.h, aiVoxelRender.cpp,
   iActionDecorator.h).
 - wTurnBased.h includes wInterface.h: CCmdCheat is used through CDynamicCast
   including its members, so it must be complete. Every includer already pulled
   it in first; this only makes the header self-contained.
 - PlatformCompat gains the BMP structures used for screenshots -- packed, with
   static_asserts pinning them to the 14/40 bytes that go to disk -- plus
   <cwctype> for iswalnum/iswpunct/iswgraph and GetDoubleClickTime.

The default (renderer-less) build still builds clean.

+++ Main/GDecal.h
@@ -22,7 +22,7 @@ struct SDecalTargetPart
 };
 struct SDecalTargetPartHash
 {
-	int operator()( const SDecalTargetPart &p ) const { return (int)p.pUser.GetPtr() ^ p.nUserID;}
+	int operator()( const SDecalTargetPart &p ) const { return (int)(intptr_t)p.pUser.GetPtr() ^ p.nUserID;}
 };
 struct SSrcPosInfo
 {
@@ -38,7 +38,7 @@ struct SSrcPosInfo
 };
 struct SSrcPosInfoHash
 {
-	int operator()( const SSrcPosInfo &p ) const { return (int)p.pUser.GetPtr() ^ p.nUserID ^ (int)p.pSource.GetPtr(); }
+	int operator()( const SSrcPosInfo &p ) const { return (int)(intptr_t)p.pUser.GetPtr() ^ p.nUserID ^ (int)(intptr_t)p.pSource.GetPtr(); }
 };
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 class CNonePart;
