+++ Misc/basic2.h
@@ -197,7 +197,7 @@ struct SPtrTest
 struct SPtrHash
 {
 	template <class T,class T1> 
-		inline int operator()( const CPtrBase<T,T1> &a ) const { return (int)a.GetBarePtr(); }
+		inline int operator()( const CPtrBase<T,T1> &a ) const { return (int)(intptr_t)a.GetBarePtr(); }
 };
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 // walks container of pointers and erases references on invalid entries
